Phone: Disable accelerometer sensor while in-call and screen UI is off
During an active ongoing call, on targets that support proximity
sensors, there is an approximate spike in the current waveform in the
order of 1 Hz. Krait is woken up frequently by the sensors even when
the Screen UI is off while in-call. This fix ensures that by disabling
the listener in such a scenario (In-Call + Screen UI turned off) there
is no spike in power consumption.

Change-Id: Ida45e1270c3f2662365fe0126a584621267cd248
CRs-Fixed: 336244, 435935